Shipit Dashboard: Move workflow in separate folder as bot. Supports Taskcluster secrets.
