+++ src/shipit_dashboard/shipit_dashboard/helpers.py
@@ -2,17 +2,9 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import json
 import hashlib
 
 
-def compute_dict_hash(data):
-    """
-    Compute SHA1 hash of a Python dict
-    """
-    data_json = json.dumps(data, sort_keys=True, ensure_ascii=True)
-    h = hashlib.sha1(data_json.encode('utf-8'))
-    return h.hexdigest()
 
 
 def gravatar(email):
